Split Classy class functions into files
