+++ test/ClassyResource_spec.js
@@ -3,7 +3,7 @@ import nock from 'nock';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 
-import Classy from '../src/Classy';
+import Classy from '../src/Classy/main';
 import ClassyResource from '../src/ClassyResource/main';
 import resources from '../src/resources';
 import _ from 'lodash';
